requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
bidict==0.22.1
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
decorator==5.1.1
elevenlabs==0.2.26
executing==1.2.0
Flask==2.3.3
Flask-SocketIO==5.3.6
frozenlist==1.4.0
h11==0.14.0
idna==3.4
ipython==8.15.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
multidict==6.0.4
openai==0.28.1
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.39
pure-eval==0.2.2
pydantic==2.4.1
pydantic_core==2.10.1
Pygments==2.16.1
python-dotenv==1.0.0
python-engineio==4.7.1
python-socketio==5.9.0
requests==2.31.0
simple-websocket==0.10.1
six==1.16.0
stack-data==0.6.2
tqdm==4.66.1
traitlets==5.10.1
typing_extensions==4.8.0
urllib3==2.0.5
wcwidth==0.2.6
websockets==11.0.3
Werkzeug==2.3.7
wsproto==1.2.0
yarl==1.9.2
